# backend/docs/swagger.yaml
definitions:
...
  github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BuildingBulkRequest:
    properties:
      ids:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - ids
    type: object
  github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BuildingBulkResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BuildingResponse'
        type: array
    type: object
  github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BuildingListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BuildingResponse'
        type: array
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BuildingResponse:
    properties:
      building_group:
        type: integer
      created_at:
        type: string
      id:
        type: string
      iws_code:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.CreateBuildingRequest:
    properties:
      building_group:
        type: integer
      iws_code:
        type: string
    required:
    - building_group
    - iws_code
    type: object
...
  github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse:
    properties:
      error:
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
...
  github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.UpdateBuildingRequest:
    properties:
      building_group:
        type: integer
      iws_code:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/facility/buildings:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BuildingListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
      summary: List buildings with pagination
      tags:
      - facility-buildings
    post:
      consumes:
      - application/json
      parameters:
      - description: Building data
        in: body
        name: building
        required: true
        schema:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.CreateBuildingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BuildingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
      summary: Create a new building
      tags:
      - facility-buildings
  /api/v1/facility/buildings/{id}:
    delete:
      parameters:
      - description: Building ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
      summary: Delete a building
      tags:
      - facility-buildings
    get:
      parameters:
      - description: Building ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BuildingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
      summary: Get a building by ID
      tags:
      - facility-buildings
    put:
      consumes:
      - application/json
      parameters:
      - description: Building ID
        in: path
        name: id
        required: true
        type: string
      - description: Building data
        in: body
        name: building
        required: true
        schema:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.UpdateBuildingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BuildingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
      summary: Update a building
      tags:
      - facility-buildings
  /api/v1/facility/buildings/bulk:
    post:
      consumes:
      - application/json
      parameters:
      - description: Building IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BuildingBulkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BuildingBulkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
      summary: Get multiple buildings by IDs
      tags:
      - facility-buildings
...
swagger: "2.0"